requirements.txt
streamlit
python-dotenv
openai>=1.0.0
selenium>=4.0.0
webdriver-manager
beautifulsoup4
